orchestrator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
redis==5.0.1
httpx==0.26.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
openai==1.10.0
anthropic==0.18.1
tiktoken==0.5.2
aiofiles==23.2.1
serpapi==0.1.5
tavily-python==0.3.2